move useNotificationInfo to hooks folder
